test: add tests for container stateful pod
