+++ tests/integration/container_runtime_args/test_container_get_args.py
@@ -1,6 +1,3 @@
-import os
-import time
-
 import pytest
 
 from jina import Client, Document, DocumentArray, Flow
