+++ tests/integration/stateful/stateful_no_snapshot_exec/Dockerfile
@@ -1,0 +1,5 @@
+FROM jinaai/jina:test-pip
+
+ADD executor.py executor.yml ./
+
+ENTRYPOINT ["jina", "executor", "--uses", "executor.yml"]
