+++ tests/integration/stateful/stateful_no_snapshot_exec/executor.yml
@@ -1,0 +1,3 @@
+!MyStateExecutorNoSnapshot
+py_modules: executor.py
+
