+++ tests/integration/stateful/stateful_snapshot_exec/executor.yml
@@ -1,0 +1,3 @@
+!MyStateExecutor
+py_modules: executor.py
+
